# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: G-Rump API
  version: 2.0.0
  description: |
    G-Rump is an AI-powered development platform that helps developers build software faster.
    
    ## Authentication
    Most endpoints require authentication via API key or OAuth token. Include the key in the 
    `Authorization` header as `Bearer <token>`.
    
    ## Rate Limiting
    API calls are rate-limited. Check the `X-RateLimit-Remaining` header for your remaining quota.
    
    ## Error Handling
    All endpoints return standard error responses with the following structure:
    ```json
    {
      "error": "Error message",
      "code": "ERROR_CODE",
      "details": {}
    }
    ```
  contact:
    name: G-Rump Support
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.grump.dev
    description: Production server
...
paths:
...
  /api/skills:
    get:
      tags: [Skills]
      summary: List all available skills
      operationId: getSkills
      parameters:
        - name: category
          in: query
          schema:
            type: string
          description: Filter skills by category
      responses:
        '200':
          description: List of skills
          content:
            application/json:
              schema:
                type: object
                properties:
                  skills:
                    type: array
                    items:
                      $ref: '#/components/schemas/SkillSummary'
  /api/skills/{id}:
    get:
      tags: [Skills]
      summary: Get skill details
      operationId: getSkillById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Skill details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillDetail'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================================================
  # Memory Endpoints
  # ============================================================================
...
  /api/integrations-v2:
    get:
      tags: [Integrations]
      summary: List all integrations
      operationId: getIntegrations
      responses:
        '200':
          description: List of integrations
          content:
            application/json:
              schema:
                type: object
                properties:
                  integrations:
                    type: array
                    items:
                      $ref: '#/components/schemas/Integration'
    post:
      tags: [Integrations]
      summary: Create a new integration
      operationId: createIntegration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIntegrationRequest'
      responses:
        '201':
          description: Integration created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Integration'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/integrations-v2/{provider}:
    delete:
      tags: [Integrations]
      summary: Delete an integration
      operationId: deleteIntegration
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/IntegrationProvider'
      responses:
        '200':
          description: Integration deleted
        '404':
          $ref: '#/components/responses/NotFound'
  /api/integrations-v2/{provider}/api-key:
    post:
      tags: [Integrations]
      summary: Store API key for integration
      operationId: storeApiKey
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/IntegrationProvider'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [apiKey]
              properties:
                apiKey:
                  type: string
      responses:
        '200':
          description: API key stored
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/integrations-v2/audit-logs:
    get:
      tags: [Integrations]
      summary: Get audit logs
      operationId: getAuditLogs
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [integration, system, security, automation, billing, agent]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Audit log entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  logs:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditLogEntry'

  # ============================================================================
  # Approvals Endpoints
  # ============================================================================
...
  /api/advanced-ai/holographic/store:
    post:
      tags: [Advanced AI]
      summary: Store in holographic memory
      operationId: storeHolographic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [key, value]
              properties:
                key:
                  type: string
                value:
                  type: string
                memoryId:
                  type: string
      responses:
        '200':
          description: Stored successfully
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
  schemas:
...
    SkillSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        version:
          type: string
        description:
          type: string
        category:
          type: string
        icon:
          type: string
        tags:
          type: array
          items:
            type: string
    SkillDetail:
      allOf:
        - $ref: '#/components/schemas/SkillSummary'
        - type: object
          properties:
            tools:
              type: array
              items:
                type: string
            hasRoutes:
              type: boolean
            hasPrompts:
              type: boolean

    # Memory Schemas
...
    IntegrationProvider:
      type: string
      enum:
        - discord
        - slack
        - spotify
        - obsidian
        - gmail
        - google_calendar
        - notion
        - twitter
        - github
        - linear
        - figma
        - home_assistant
        - elevenlabs
        - twilio
        - sendgrid
        - stripe
        - custom
    Integration:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        provider:
          $ref: '#/components/schemas/IntegrationProvider'
        status:
          type: string
          enum: [active, disabled, error, pending]
        displayName:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateIntegrationRequest:
      type: object
      required: [provider]
      properties:
        provider:
          $ref: '#/components/schemas/IntegrationProvider'
        displayName:
          type: string
    AuditLogEntry:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        actor:
          type: string
        action:
          type: string
        category:
          type: string
          enum: [integration, system, security, automation, billing, agent]
        target:
          type: string
        metadata:
          type: object
        createdAt:
          type: string
          format: date-time

    # Approval Schemas
...
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              code:
                type: string
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
security:
  - bearerAuth: []
  - apiKey: []